chore: Removed eslint
a new client WorkerLayoutClient component for the
workers route that provides a split view with a worker-side list and
detail pane. Implement a useWorker query hook to fetch worker data,
wire up loading state and header actions (settings menu + delete
action), and render WorkerThreadsList and WorkerDeleteModal.

Remove several empty eslint config/ignore files from infra apps
(cloud-primary, workingparty, remote-browser, pulumi-backend) and
clean up tsconfig includes to avoid referencing removed files.

Also remove ESLint-related devDependencies and a lint script from
infra/doprocess package.json to reflect consolidated linting setup.

These changes add the interactive client UI for worker pages and
clean up obsolete ESLint config and package entries.
