--- infra/.vscode/settings.json
@@ -1,7 +1,0 @@
-{
-  "eslint.workingDirectories": [
-    {
-      "mode": "auto"
-    }
-  ]
-}
