--- infra/apps/cloud-primary/.eslintignore
@@ -1,2 +1,0 @@
-dist
-.turbo
