--- infra/apps/doprocess/.eslintignore
@@ -1,2 +1,0 @@
-dist
-.turbo
