+++ infra/apps/remote-browser/package.json
@@ -3,18 +3,10 @@
     "private": true,
     "license": "MIT",
     "main": "src/index.ts",
-    "scripts": {
-        "lint": "eslint ./src/**/*.ts"
-    },
+    "scripts": {},
     "devDependencies": {
         "@infra/pulumi": "workspace:*",
         "@types/node": "22.15.18",
-        "@typescript-eslint/eslint-plugin": "8.30.1",
-        "@typescript-eslint/parser": "8.30.1",
-        "eslint": "8.57.1",
-        "eslint-plugin-import": "2.31.0",
-        "eslint-plugin-n": "17.17.0",
-        "eslint-plugin-promise": "7.2.1",
         "tsup": "8.4.0",
         "typescript": "5.8.3"
     },
