--- infra/apps/workingparty/.eslintignore
@@ -1,2 +1,0 @@
-dist
-.turbo
