--- infra/packages/eslint-config/README.md
@@ -1,3 +1,0 @@
-# `@turbo/eslint-config`
-
-Collection of internal eslint configurations.
