--- infra/packages/eslint-config/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@infra/eslint-config",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "files": [
-    "library.js"
-  ],
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "8.30.1",
-    "@typescript-eslint/parser": "8.30.1",
-    "eslint-config-prettier": "10.1.8",
-    "eslint-config-turbo": "2.5.8",
-    "eslint-import-resolver-typescript": "4.4.4",
-    "eslint-plugin-import": "2.31.0",
-    "eslint-plugin-only-warn": "1.1.0",
-    "typescript": "5.8.3"
-  }
-}
