+++ infra/packages/pulumi/package.json
@@ -3,9 +3,7 @@
     "private": true,
     "license": "MIT",
     "type": "module",
-    "scripts": {
-        "lint": "eslint ./src/**/*.ts"
-    },
+    "scripts": {},
     "exports": {
         "./azure": {
             "import": "./src/azure/index.ts"
@@ -24,15 +22,8 @@
         }
     },
     "devDependencies": {
-        "@infra/eslint-config": "workspace:*",
         "@infra/typescript-config": "workspace:*",
         "@types/node": "22.15.18",
-        "@typescript-eslint/eslint-plugin": "8.30.1",
-        "@typescript-eslint/parser": "8.30.1",
-        "eslint": "8.57.1",
-        "eslint-plugin-import": "2.31.0",
-        "eslint-plugin-n": "17.17.0",
-        "eslint-plugin-promise": "7.2.1",
         "fflate": "0.8.2",
         "glob": "11.0.3",
         "typescript": "5.8.3"
