+++ web/apps/workingparty/src/hooks/data/account/useAccount.ts
@@ -1,4 +1,4 @@
-import { UseQueryResult, useQuery } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 
 export type Account = {
     id: string;
@@ -6,15 +6,13 @@ export type Account = {
     createdAt: number;
 };
 
-export function useAccount(accountId: string | undefined): UseQueryResult<Account, Error> {
+export function useAccount(accountId: string | undefined) {
     return useQuery({
         queryKey: ['accounts', accountId],
         queryFn: async () => {
             const response = await fetch(`/api/accounts/${accountId}`);
             if (response.status < 200 || response.status > 299) {
-                return {
-                    isLoggedIn: true
-                };
+                throw new Error(`Failed to fetch account data: ${response.statusText}`);
             }
 
             const userData = await response.json() as Account;
