+++ web/packages/email-templates/package.json
